Hospital-Management-API/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
Markdown==3.7
openai==1.40.6
openpyxl==3.1.5
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.5
typing_extensions==4.12.2
